[SYCL] Fix vec class alignment on windows platform (#4953)
Currently the sycl::vec type can be copied in the way which doesn't
preserve the default alignment on windows. This can causes crashes
since the sycl:;vec code expects the vector to be aligned and uses
vector instructions. We used default alignment because we cannot
set correct alignment in all cases. The patch adds alignment of
vector types, if alignment required is larger than 64, it is limited to 64.