Add --spirv-tools-dis option to disassemble using SPIRV-Tools
llvm-spirv's text rendering of a SPIR-V module uses an internal
non-standard format that is not compatible with any other tools.

Add support for emitting SPIR-V assembly using the more widely used
SPIR-V Tools format.  This is achieved by linking to SPIRV-Tools and
invoking its disassembler.  For now, this functionality is only
available if the SPIRV-Tools package is found at build configuration
time.

The main motivation for this change is the proposed integration of
llvm-spirv in Clang.  Particularly, we would like Clang invocations
such as `clang -S -target spirv` to produce the standard SPIR-V
assembly format instead of llvm-spirv's internal format.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/beb88bf