[SYCL] Use incude/sycl/CL as the target to check copied headers
When the sycl-headers target calls the copy_directory command, the
command sees that the ${OUT_HEADERS_IN_CL_DIR} output already exists
and the files are already there. Before the patch, the command
looked for the include/CL directory, not include/sycl/CL one, there was
no such directory and the headers were copied again and again.

Now when we ask CMake to build the sycl-headers target twice, the
'Copying SYCL headers ...' message is displayed only after the first
time.

Co-authored-by: tcreech-intel <timothy.m.creech@intel.com>