[SYCL] accessor::get_pointer() should always return base pointer, even for offset accessors (#4687)
According to the SYCL 2020 spec get_pointer() returns a pointer to the beginning of the underlying buffer, even if it is offset. Presently our get_pointer() implementation returns an offset pointer when using an offset accessor. This PR fixes this. 

The Jenkins Summary is failing because the CTS tests have not been updated for this requirement.  I have a PR introducing this update on the CTS used by the CI system.

Signed-off-by: Chris Perkins <chris.perkins@intel.com>