[SYCL][CUDA] Update program manager and queue to resolve multi-targeting issues (#4921)
This PR makes two changes, the first is it moves the macro which prevents `__devicelib_assert_read` being used for `nvptx64` devices. This is done to resolve an issue where the binary images of `spirv64` and `nvptx64` are neither identical nor disjoint (have no kernels in common). The program manager needs binary images to be identical or disjoint, it fails otherwise. This creates a kernel of the same name as when building for `spirv64` but it does not use `__devicelib_assert_read`. 
The second it prevents errors being thrown in the program manager when the binaries compatibility check returns false. This is to allow for multi-targeting to be used with module splitting. 
A cuda and hip only regression test is added to check for successful compilation with multi-targeting and module splitting options.