[SYCL][LIBCLC] Create remangled names aliases (#5041)
This PR creates an alias of the libclc remangled function, so that permutations of the remangled function are not removed. 
For example with long 32 bit, signed char remangling  `f(long, char)` will be remangled to `f(long long, signed char)`. The alias of the original name, for l32 & signed char, becomes `f(int32, signed char)`. Therefore no `f(long long, char)` is declared. This PR solves the issue by creating an alias of the remangled function which is the original function transformed by all parameter type replacements that are not aliases. As a result `f(long long, char)` becomes an aliase of `f(long long, signed char)`.