Change the OpenCL extension
After discussions with @bashbaug, @GarveyJoe, and @artemrad; we decided
to change the OpenCL extension to avoid returning a pointer to the
global variable.  Instead, we add two new APIs that can read or write
the content of the variable.  This avoids the need to precisely define
the semantics of the returned pointer.