[SYCL] Add element size argument to piKernelSetArg
This patch comes from an attempt to fix #5007.

The issue there is that for local kernel argument the CUDA plugin uses
CUDA dynamic shared memory, which gives us a single chunk of shared
memory to work with.

The CUDA plugin then lays out all the local kernel arguments
consecutively in this single chunk of memory.

And this can cause issues because simply laying the arguments out one
after the other can result in misaligned arguments. In #5007 for example
there is an `int` argument followed by a `double4` argument, so the
`double4` argument ends up with the wrong alignment, only being aligned
on a 4 bytes boundary following from the `int`.

It is possible to adjust this and fixup the alignment when laying out
the local kernel arguments in the CUDA plugin, however before this patch
the only information in the plugin would be the total size of local
memory required for the given arguments, which doesn't tell us anything
about the required alignment.

So this patch propagates the size of the elements inside of the
local accessor all the way down to the PI plugin through
`piKernelSetArg`, and tweaks the local argument layout in the CUDA
plugin to use the type size as alignment for local kernel arguments.