[SYCL][ESIMD] Add test proxy for move constructors
Because copy constructor works as the default fallback for every case with move
constructor disabled or not provided, there is no way for a caller to
differentiate between actual call to move constructor and copy constructor.
Same goes for copy and move assignment operators.

Therefore a test proxy is the only way left to check the move constructor and
the move assignment works as expected.

Signed-off-by: Yuriy Kochetkov <yuriyx.kochetkov@intel.com>