[SYCL] Fixes memory dependency leaks caused by failed kernel enqueue
If a kernel enqueue fails the runtime will immediately try and clean it
up. However, if it has any dependencies or users the cleanup will be
skipped. This can cause the dependencies to stay alive and leak.
These changes forces a full sub-graph cleanup of the command if
enqueuing failed. Additionally, sub-graph cleanup is changed to account
for failed kernel enqueues and will remove the failed command from its
leaves.

Signed-off-by: Steffen Larsen <steffen.larsen@intel.com>