[SYCL] Fix memory leak in reduction resources
Reductions that require additional resources, such as buffers, can
currently create a circular dependency between the resources and the
commands issued by the reductions. These changes clear up this
dependence in a similar way to how streams are transferred by
transferring ownership of the resources to the commands and ensuring
release when cleaning up the commands.

Signed-off-by: Steffen Larsen <steffen.larsen@intel.com>