[SYCL][HIP] Fix max constant memory device query
This fixes the `Basic/info.cpp` test for HIP AMD.

The issue here is that AMD GPU report a very large constant memory
(confirmed using `clinfo` as well). But the HIP entry point to query
that information takes a `int*`. So the value will show up as negative,
even though it's the correct value.

So remove the assertion on if the value is positive and cast it back to
`unsigned` before returning it.