[Driver][SYCL] Disable the warning when fsycl-host-compiler is used (#5124)
The warning is emitted because we aren't doing any host compilation at that point. But the warning can be misleading to the users, they may think there is something inappropriate with usage of -fsycl-host-compiler. So instead of asking user not to use the option we are just ignoring the warning.