[SYCL] Fix kernel bundles don't really carry kernel IDs (#5121)
Fix that interop kernel bundles don't really carry kernel IDs for contained kernels. This change adds In enqueueImpKernel another condition to taking the path using the associated kernel_bundle, namely that interop bundles can't be used in the first branch. (This is because the kernels in interop kernel bundles (if any) do not have kernel_id and can therefore not be looked up, but since they are self-contained they can simply be launched directly.)