[SYCL] Remove circular dependency between piEventRelease and cleanup()
Event is retained explicitely when created in createEventAndAssociateQueue()
to ensure that the event is not destroyed before it is really signalled.
Ref count is decremented in Event->cleanup() after event is singalled. This
means that if reference count for an event is zero then cleanup was already
called for this event and should not be called anymore.

Interop events and user events don't have associated command list, queue,
kernel and waitlist, so we don't need to call cleanup() for interop and user
events.