[SYCL] Remove branch which is never taken
Event is retained explicitely when created to ensure that the event is
not destroyed before it is really signalled. It is release in
Event->cleanup after event is singalled. This means that if reference
count for an event is zero then cleanup was already called for this
event and is not going to be called anymore.