[SYCL][HIP] Remove arch requirement for running lit tests
This flag was introduced before all the tests that run code were moved
to the `llvm-test-suite` repository. The tests left in this repository
are only compiler tests so the offload architecture can simply be
hardcoded as the tests don't depend on what hardware is available.

This simplifies building and running the tests for hip.