[SYCL] Diagnose unresolved symbols in L0 modules (#5267)
Diagnose an error in the Level Zero version of `piProgramBuild` and
`piProgramLink` if the built program has any unresolved symbols.  Also,
include an error message with a list of the unresolved symbols in the
build log, which is incorporated into the `what` string of the
exception that the runtime throws.  As a result, the user will see a
list of the unresolved symbols when the exception terminates the
application.

Previously, no error was diagnosed from `piProgramBuild` or
`piProgramLink` when there were unresolved symbols, but the first call
to `piKernelCreate` would fail (without any indication about the cause
of the failure).