[SYCL][CUDA] Remove unsued _pi_kernel constructor (#5255)
This constructor is not used anywhere.

This resolves the discussion on intel/llvm#5152 about how
`cuda_piKernelGetGroupInfo` would be called twice when this constructor
is used, as it defers to the other constructor which also calls it. But
this constructor wasn't being used anyway so it can simply be removed.