[SYCL] Partial fix for stream problem (#5136)
If work group size wasn't set explicitly then buffer for the stream object will be created with zero size. WG size is received later from work group size kernel attribute if set, or set to the default size with size = 1 for every dimension. Right now we can't get this attribute, it will be possible when the feature of containing kernel attributes by device image will be implemented.
So, when work group size is needed for stream, we suppose that its size is default for the correct working of stream in the cases when the attribute isn't set.
Test for this fix : https://github.com/intel/llvm-test-suite/pull/617
Signed-off-by: mdimakov <maxim.dimakov@intel.com>