[Driver][SYCL] Enable generation of spv based fat objects (#4608)
When generating fat objects with -fsycl, our default behavior is to
generate and store the LLVM-IR for the device side compilation.  Update
the behavior, introducing -fsycl-device-obj=arg (where arg is spirv or
llvmir) to create the desired 'object' contents.

This is a precursor to the enabling the ability to consume these spv
based objects (or llvm-ir) seamlessly.