[SYCL][HIP][libclc] Fix wrong address spaces for HIP (#4966)
Fixes wrong address spaces for `event_t`. It should be private, but in SYCL it was defaulting to generic. On the other hand in libclc pointer to `event_t` was private, while it should be generic. (We don't see similar issues on NVPTX, as there private address space is in fact implemented as generic in IR).

Change to test suite: https://github.com/intel/llvm-test-suite/pull/570