[SYCL] Improve mutation of literal address space for variadic printf
The initial implementation of address space mutation assumed that in
presence of `experimental::printf` wrappers at the IR level, each
wrapper would use its dedicated signature of the `__spirv_ocl_printf`
builtin. However, this is not the case when the variadic signature of
the builtin is employed - in that scenario, the wrapper does not
become variadic, so `__spirv_ocl_printf` may have multiple wrapper
instances using it.

Adjust the algorithm for removing generic AS functions accordingly and
add LIT coverage for the non-inline variadic case.

Signed-off-by: Artem Gindinson <artem.gindinson@intel.com>