[SYCL] Filter implicit kernel bundle images
The runtime currently uses implicit kernel bundles to allow for
operations such as setting specialization constants. Since the handler
does not know the kernel to launch before it is being launched, the
implicit kernel bundle created is not limited to specific kernels.
As an effect of this, the kernel bundle may try to build more device
images than it needs.
These changes make the runtime filter the device images of implicit
kernel bundles as soon as the relevant kernel is known.

Signed-off-by: Steffen Larsen <steffen.larsen@intel.com>