[SYCL][CUDA][HIP] Fix PI_MEM_ALLOC_DEVICE USM pointer query (#5325)
This fixes the `USM/pointer_query.cpp` test from `llvm-test-suite` for
HIP.

For HIP the code was using the device pointer to index the device array
for some reason, which is obviously incorrect, now it should be using
the correct property to get the device index on the hardware.

In addition for both HIP and CUDA the query was using the device index
to index into the array of devices on the platform, this is incorrect
since #4571 which moved every device in its own platform. Therefore to
get the correct device we now need to use the index on the list of
platforms, and then `0` on the list of devices.