[Driver][SYCL] Improve archive/listfile representation in the toolchain
When dealing with archives and SYCL device compilations, we recently made
adjustments to expect SPIR-V based fat objects.  These are unbundled as
list files so we can individually convert the SPIR-V to LLVM-IR as needed.

The output from the unbundling was being represented as archives instead of
the true form of temporary list files.  This caused problems when having
multiple target output when one of the targets was not spir64 based.

Update representation to use the proper file type being created from the
unbundling of the archives.