[SYCL] Remove doc [[intel::reqd_work_group_size]] (#5387)
This extension document describes an old form of the C++ attribute
`[[intel::reqd_work_group_size]]`, which is in the `intel` namespace.
This attribute has been adopted into SYCL 2020 in the `sycl` namespace,
and the compiler does not even recognize the form in the `intel`
namespace.  There is, therefore, no need to keep this old extension
document.