[sycl-post-link] Check if the tool met at least one device global variable
If no device global variable has been met, the [SYCL/device globals] property
set should not be written into the property file. The check

if (!DevGlobalPropertyMap.empty())

has became redundant and was removed because the ImgPSInfo.DeviceGlobalsMet
variable will be set to true if and only if there is at least one device global
variable in the LLVM IR module.