[sycl-post-link] Implement the DeviceGlobals module pass
The pass enriches the LLVM IR with the !spirv.Decorations metadata
for the properties of every device global variable.

We suppose the enumeration orders in the host_access and init_mode
enumerations in the SYCL headers are the same as in the descriptions of
the HostAccessINTEL and InitModeINTEL decorations in the
SPV_INTEL_global_variable_decorations extension.