[Driver][SYCL] Initial support for allowing fat static -lname processing
When processing for fat static libraries, add searches for -lname type
libraries passed on the command line.  This also takes advantage of
the -Ldir options and LIBRARY_PATH environment variable for locating
the libraries.

When a library is found (with a .a extension) that is what is used for
unbundling and added to the device link.

NOTE:  This behavior does not match the recent implementation done for OpenMP
offloading with fat static libraries.  That implementation is directly
integrated into the llvm-link call.  It would be good to also follow this
for consistency, but will require some refactoring to accomplish.