Fix dead ptr usage if leaf buffer overflows
When leaf buffer overflows we take oldLeaf and make a new one dependent on it.
In case of connect command insertion we again should take oldLeaf and do the same. So we do recursive work with the same buffer.
This change helps to eliminate it by not adding connect command to the leaf buffer since we already put to the buffer or will put command depending on connectCmd.
It also tries to remove duplicated dependencies between newLeaf and oldLeaf. If connection command inserted new leaf already depends on old leaf via connect cmd. So we do not put old leaf as standalone dependency.

Signed-off-by: Tikhomirova, Kseniya <kseniya.tikhomirova@intel.com>