[sycl-post-link] Extract the CompileTimePropertiesPass pass
The pass provides the lowering of registered in the pass compile-time
properties into metadata nodes with SPIR-V decorations as it is
described in the design document [1]. Only properties on global
variables are supported yet.

Some properties require special handling, for example the corresponding
SPIR-V decoration requires two "extra operands". For example, the
"host_access" property of a device global variable.

[1] https://github.com/intel/llvm/blob/sycl/sycl/doc/CompileTimeProperties.md#property-on-a-global-variable