[SYCL][DOC] Move supported extension docs (#5421)
Move the supported extensions to the "sycl/doc/extensions/supported"
directory, and add a README explaining the content of that directory.
Each extension document has also been renamed to match the name of
its feature-test macro.

Links to these extensions have been removed from the table in
"sycl/doc/extensions/README.md" because we want to avoid the
possibility of that table getting out-of-date with the directory
contents. Instead, users can see the list of supported extensions
by listing the content of the directory.