[SYCL][DOC] Move proposed extension docs (#5437)
Move the proposed extensions to the "sycl/doc/extensions/proposed"
directory, and add a README explaining the content of that directory.
Each extension document has also been renamed to match the name of its
feature-test macro.

Links to these extensions have been removed from the table in
"sycl/doc/extensions/README.md" because we want to avoid the
possibility of that table getting out-of-date with the directory
contents.  Instead, users can see the list of proposed extensions
by listing the content of the directory.