[SYCL][DOC] Move deprecated extensions (#5458)
Move these deprecated extension specifications to the "deprecated"
directory, but add a signpost file in the "supported" directory pointing
at the deprecated extension.

We will follow this convention when a supported extension is deprecated.
The deprecated extension will move to the "deprecated" directory, and
we will change the file in the "supported" directory into a signpost.  Thus,
any external links to the spec will still point to something that leads to the
deprecated extension.