[SYCL][ESIMD] Update vc-intrinsics deps to the top of the trunk (#5460)
dpcpp_staging branch is supposed to be used to workaround issues faced
due to async pulldowns from llvm-project upstream to vc-intrinsics project and
to DPC++ project. As soon as all 3 projects are in sync we should switch
DPC++ to the main branch of vc-intrinsics. This is what this change is
doing. Meanwhile dpcpp_staging will continue to be used to workaround next
set of upstream pulldown issues.