[SYCL] Fix performance regression in parallel_for with using id class (#5385)
In https://github.com/intel/llvm/pull/5118 conversion for user defined types which are implicitly converted from sycl::items was added. It caused regression in case when sycl::id type is used, because with this change it converts to sycl::item but in fact it is not needed. That is why sycl::id class case should be checked explicitly.