[SYCL][Doc] Add sub_group::load/store extension
The original extension document containing the specifications for the
sub_group::load and sub_group::store functions was previously removed,
leaving these extensions undocumented.

Although we plan to deprecate and replace these extensions in the near
term, providing no documentation is not user-friendly.

Note that the documentation introduced by this PR describes only the
functionality that is already implemented in the sub_group class.

Signed-off-by: John Pennycook <john.pennycook@intel.com>