[SYCL] Implement initial device_global registration (#5499)
These changes add required interface for the integration footer to register the `device_global` of a program as well as reading the
extended information supplied by sycl-post-link through the "SYCL/device globals" property. The information is stored in a map
managed by the program manager.

Note that this is the first change to the runtime relating to `device_global` and as such the feature is not complete nor fully testable. It simply intents to add an interface for the integration footer as well as a consumer for the information sycl-post-link will be generating when future work is added.

See [sycl/doc/DeviceGlobal.md](https://github.com/intel/llvm/blob/sycl/sycl/doc/DeviceGlobal.md).