[sycl-post-link] Fix a crash during spec-constant properties generation
The sycl-post-link tool crashed during processing an IR where a
specialization constant with a padding is present, for example a
specialization constant of such type:

struct alignas(32) coeff_str_aligned_t {
  std::array<float, 3> coeffs;
  size_t number;
};

In the IR, the constant looks like the following:

@_ZL8coeff_id = internal addrspace(1) constant %"class.cl::sycl::specialization_id" { %struct.coeff_str_aligned_t { %"class.std::array" zeroinitializer, i64 0, [8 x i8] undef } }, align 32

That '[8 x i8] undef' led to the crash.

This patch fixes the issue.

Co-authored-by: Steffen Larsen <steffen.larsen@intel.com>
Signed-off-by: Pavel Samolysov <pavel.samolysov@intel.com>