[SYCL] Avoid strict aliasing violation when reading from byte arrays
When reading properties produced by sycl-post-link, the runtime will
cast a byte-array pointer to another pointer of a diffrent type and
read from the resulting pointer. This is a violation of the strict
aliasing rule and as such is UB. These changes replace the violating
reads with memcpys from the byte arrays.

Signed-off-by: Steffen Larsen <steffen.larsen@intel.com>