[sycl-post-link] Optimize __spirv_SpecConstantComposite generation
Instead of generating a '__spirv_SpecConstant' invocation for every
element of a composite data type with value 'undef' and then combining
thir results into a call to '__spirv_SpecConstantComposite', the
following optimization has been implemented. If the default value of a
specialization constant or a part of such default value is undefined,
the value (or its part) is passed into an invocation of the
'__spirv_SpecConstantComposite' function as is.