[SYCL] Fix SYCL Kernel Body Check (#5546)
SYCLKernelAttr can be used to check if a given function is
a SYCL kernel body function or not. A SYCL kernel body function
is the operator() method of a SYCL Kernel Functor or Lambda Function.
The existing check which only tested if a FunctionDecl was an
operator or not, allowed for false positives. This patch adds an
additional check for SYCLKernelAttr, which will be attached to
the kernel body function's parent/ parent's parent (for wrapped
kernels)

Signed-off-by: Elizabeth Andrews <elizabeth.andrews@intel.com>