[sycl-post-link] Initialize the integer Value variable (#5585)
A static analysis tool emits a warning about uninitialized integer
variable 'Value' in the CompileTimePropertiesPass.h header. Actually,
the variable is used just to pass a reference to it as an output
parameter into the 'getAsInteger()' function and will be initialized
there, but if an error occurs, the variable will remain uninitialized.
Although there is a check for the error, let the variable be initialized
is the best strategy.