Do not generate Sphinx doc for OpenCL/SPV exts
Do not generate Sphinx documentation for the OpenCL or SPIR-V
extensions because they are written in AsciiDoc, which Spinx does not
support well.  Instead, link to their GitHub location, and readers can
read GitHub's rendering of the AsciiDoc documents.  This is the same
thing we do already for the SYCL extension specifications.