[SYCL][DOC] Change extension names to lower case (#5607)
I found the all-upper-case names difficult to read, especially in a
directory listing where all the files are upper case.  Change them all
to lower case.  The name of the extension and the name of the file are
now lower case.  The name of the feature-test macro remains upper case.