[SYCL] Implement resource pool for implementation allocations
These changes add a "resource pool" used by the SYCL implementation to
get auxiliary device memory resources. When a resource is deleted it
returns its memory to the pool to be reacquired by other internal logic.
Additionally, reductions that need additional resources are made to use
the pool when getting these.

Signed-off-by: Steffen Larsen <steffen.larsen@intel.com>