Define new buffer_location runtime property to be passed into malloc_device
-----------------------------------------------------------------------------------

This serves as a temporary solution to supporting usm buffer location, and it is the easiest to implement.
Align with changes in https://github.com/intel/llvm/pull/5634

A more complex workaround is to define a new malloc API that accept sycl::ext::oneapi::experimental::properties as its property argument of malloc.

The full solution is to define a malloc api that takes sycl::ext::oneapi::experimental::properties as property argument and returns annotated_ptr