[spirv-to-ir-wrapper] Prevent non-IR from being passed along
Introduce an option -skip-non-llvmir which can be used to prevent non-IR
files from being passed through the tool.  There are usage models involving
archives in which the 'fat archives' do not contain all fat objects.  In
the case where this occurs, provide a way to prevent the non-IR object
from the device side to be passed thorugh as it is not valid at the
device link step.