[SYCL] Fix host device local accessor alignment (#5554)
Local kernel arguments must be aligned to the type size, simply using `std::vector<char>` doesn't always provide the correct alignment. So this patch adds extra padding to the vector and ensures that the pointer returned for the accessor is actually aligned to the type size.

This issue was exposed by: https://github.com/intel/llvm-test-suite/pull/608, which was a follow up to fixing local accessor alignment for the CUDA plugin.