[SYCL] Ensure correct sycl::errc for SYCL2020 (#5582)
The SYCL2020 specification has quite a few errc
values it expects thrown exceptions to have. Most of
the existing ones are fine, but there are several cases
that need updating. After this there are some other
changes that will need to be done for errc
conformance, but those changes are slightly more
involved. They will be done in a follow up PR s.

The tests were updated as well. intel/llvm-test-suite#857
has the matching changes.