[SYCL] Fix bug with constexpr recursion (#4257)
Commit 9a9a018c823ec introduced a regression in following case - 

E.g.

```
  struct ObjInit {
    int a;
  } objbar = {constexpr_recurse(1)};
```

The assert was hit when DeclStmt contains both TypeDecl and VarDecl i.e. a CXXRecordDecl which is not constexpr and a VarDecl which is. This PR removes code added in call graph creation. This code is not technically not required since the AST visitor in SemaSYCL, will handle non-traversal of nodes representing constexpr variables.

Closes: https://github.com/intel/llvm/issues/4048

Signed-off-by: Elizabeth Andrews <elizabeth.andrews@intel.com>